API/requirements.txt
Pillow==8.4.0
keras==2.8.0
matplotlib==3.5.1
numpy>=1.20.3
pandas==1.3.4
seaborn==0.11.2
scikit-image==0.19.2
tensorflow==2.8.0
fastapi==0.75.1
uvicorn[standard]==0.17.6